chore: deduplicate code between pepr-build and pepr-build-wasm
